requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.2
django-absoluteuri==1.3.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
executing==1.2.0
fontawesomefree==6.4.0
inflection==0.5.1
ipython==8.14.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
pure-eval==0.2.2
Pygments==2.15.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
wcwidth==0.2.6
